requirements.txt
airsim
numpy
pillow
opencv-python
scikit-learn
